bash: Stop using double-quoted strings in SQL
Today I noticed that my new shell sessions were giving me fun sqlite
errors before and after each command. It turns out that I was using
double quotes for a string literal, and the SQLite 3.41.0 release[1]
just disabled a feature[2] which was permissively allowing this to work.

I'm going to be honest, I didn't realize that single quotes were
required, and I feel like I've written a decent amount of SQL. It seems
a little strange for SQLite to do this in a minor version bump, but
whatever, I guess?

[1]: https://www.sqlite.org/draft/releaselog/3_41_0.html
[2]: https://www.sqlite.org/draft/quirks.html#dblquote

Signed-off-by: Stephen Brennan <stephen@brennan.io>