Update theme switching
For the longest time, theme switching has left my dotfiles in a constant
state of having unstaged changes. There were always modifications left
behind by my sed invocations tweaking the configs in bin/light or
bin/dark.

The thing is, I have a framework for m4 templating configs. Instead of
directly editing the configs, let's just set THEME as an M4 variable and
template relevant configs. The light and dark scripts still need to
retain some ability to notify programs of changes, and some of the
configs are still a bit convoluted to support that (e.g. vim). But on
the whole, this is an incredible improvement, as it means I should
always have a clean dotfile directory.

Signed-off-by: Stephen Brennan <stephen@brennan.io>