remove tagsearch
This was a really cool bit of code! I'm sad to delete it. But the thing
is, I don't use it. It used a forked copy of ccls and I don't even use
ccls anymore. I'd love to come up with a "browse to the definition of a
symbol" tool based on a more general LSP client, but honestly, given
that I never used this one, it seems it's not a high priority.

Signed-off-by: Stephen Brennan <stephen@brennan.io>