setup: support overrides
For a while, my work dotfiles have been a fork of my home dotfiles, and
I've cherry-picked between them. But the forks have diverged, and it's
difficult to maintain them.

Introduce a new "extension" system. The idea is that you can stick a
submodule or whatever you want into the "ext" directory. We'll
automatically source the corresponding "libsetup.sh" script there, which
can append to any of the link arrays.

When setting up links, we prefer to take files from the extension
directory. The idea here is that we can completely override files that
need to be significantly different (for example, email configuration).
To avoid complicating things too much, it will be better to create new
files rather than override existing ones.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>