Dramatically improve my use of m4
I finally decided to RTFM for m4, and now that I have a slight bit more
understanding of how it works, I realize my previous approach was, maybe
not silly, but a bit over the top.

I use m4 as a portable file templating system. The best thing is to have
files look just like they would once they are templated, except for the
strings that need to get replaced. M4 can do this, but you need to
ensure that (a) none of the tokens in the file turn out to be macros,
and (b) you don't encounter your quote strings in the file. To
resolve (a), m4 has a handy -P option, which prefixes all of its builtin
macros with "m4_", something that I think I can be confident doesn't
occur in most configs. To resolve (b), I made my quote characters
something suitably obscure: @[ and ]@. While [[ and  ]] would be more
familiar to users of autotools, some of my configs actually do contain
those characters.

Anyway, changing the quote character does require a "changequote()"
directive, but we can provide that on behalf of each file using "cat" --
finally a use of cat which actually... conCATenates files.

We adopt a convention that all macros that may be substituted in configs
look like X_NAME_X, so they're visually distinctive. With that, all the
m4 files are much simpler, as we get rid of the changequote business.

I did need to rename some theme files from, e.g. solarized_light, to
solarized-light. This allowed me to use the macro "solarized-X_THEME_".
This is not required, I could use the m4 quote character, but I honestly
I prefer to rename the files. I also noticed that muttrc no longer needs
m4, so I converted it back.

Finally, to tie everything together, add a ./diffm4 script which helps
you determine whether things look like you expect.

Signed-off-by: Stephen Brennan <stephen@brennan.io>