Try to use true color support wherever possible
It seems like "alacritty-direct", "konsole-direct", and "tmux-direct"
are the most capable versions of color support for TERM, and I should be
using those ideally. The COLORTERM=true variable is another way to
signal that I have true color support, and that should sent via SSH if
possible. However, SSH servers need to configure AcceptEnv to receive
it, which is not likely.

Vim needs direct color configured explicitly, so do that.

Sadly, not everything has the most recent terminfo databases that are
aware of these newer TERM variables. I can handle that by falling back
to the most recent option in my SSH host config.

Signed-off-by: Stephen Brennan <stephen@brennan.io>